Force pre-population of IN data for correct uploading of info
Instead of taxing the API to an even greater extent, it was helpful to
simply do the work to provide the CSV with the public_ids of invoices
and payments ahead of time instead of doing a lookup of the public_id
based on a custom field or a transaction Id. I may still want to
implement that type of functionality in the future.